Add files via upload
The first three consists of the components to be used in the algorithm. Algorithm so far initializes from a file all of the components with the constructors deriving the various final boolean values. At the moment, Region needs to be adjusted to account for the boolean values part, but the rest is good for now. I plan on moving the initialization process into an initialization method for future error checking.