Add files via upload
I have successfully coded the first step of the algorithm. Changes to be noted would be the lack of an adjacencyList() method because this has been relabeled to procedure1(). This was necessary (as described in the comments) to help optimize the overall algorithm. Vertices now keep track of their degree. Additional methods for setting, incrementing, and getting variables were part of this update.