Add files via upload
An error-free update of the previous commit. Procedure 1 has been split back in the Algorithm class while the adjacencyList() method has been returned, independent of procedure1, to the Graph class. This keeps a clear separation of the Algorithm class executing procedures according to the properties, characteristics, and information provided by the Graph class.