Add files via upload
This update adds weights to vertices. There is a great deal of bookkeeping work done for this update primarily. Clarification and clarity have been added to increase readability. Lastly, procedure2() should be completed by now, and cycleResolver() added as a place holder for now. Considerations for the clone() method for a deep versus shallow copy are underway; however, the code attempts to keep clear the fact that the copies are being edited rather than the graph's copy of lists of objects.